Estimate rake weights for input survey data & population distributions

This is a thin wrapper for anesrake. If return_stats = TRUE, it returns a list with (1) a survey dataset and (2) the anesrake() object. Otherwise, it prints a summary and returns just the survey dataset with a weight variable appended.

Usage

rake_weight(
  svy,
  pop,
  idvar,
  weightvar = "weight",
  return_stats = TRUE,
  print_name = "",
  ...
)

Arguments

svy

a survey data frame

pop

a list that holds population distributions for the weighting variables.

idvar

name of varible that holds unique id

weightvar

name of output weight variable

return_stats

If TRUE, the function will instead return a list with 2 elements: (1) the survey data frame with a new weight variable, and (2) the object returned directly by anesrake(). This is helpful if you want to later pull elements such as design effect for confidence interval calculations.

print_name

header to print in summary (useful for log output)

...

other arguments passed to anesrake()

Examples

data(svy, pop)
svy <- svy$person

# determine target weight values
# the population is determined by a survey dataset which is itself weighted
wtvars <- c("age_weight", "income_weight", "race_weight")
pop_target <- sapply(wtvars, function(x) weights::wpct(pop[[x]], pop$stwt))

# compare distributions for survey and population
pop_target
#> $age_weight
#>     18-34     35-54       55+ 
#> 0.3514503 0.3486601 0.2998896 
#> 
#> $income_weight
#>      0-25K     25-35K     35-50K     50-75K    75-100K   100-150K      150K+ 
#> 0.17353726 0.09405487 0.14064324 0.18922885 0.15252025 0.16103245 0.08898309 
#> 
#> $race_weight
#>                 White              Hispanic Not white or Hispanic 
#>            0.73559778            0.17719770            0.08720452 
#> 
sapply(wtvars, function(x) weights::wpct(svy[[x]]))
#> $age_weight
#>     18-34     35-54       55+ 
#> 0.2766497 0.2944162 0.4289340 
#> 
#> $income_weight
#>      0-25K     25-35K     35-50K     50-75K    75-100K   100-150K      150K+ 
#> 0.15071973 0.12785775 0.14309907 0.20491109 0.15071973 0.15410669 0.06858594 
#> 
#> $race_weight
#>                 White              Hispanic Not white or Hispanic 
#>            0.82387807            0.09652837            0.07959356 
#> 

# run weighting
svy_wts <- rake_weight(svy, pop_target, "Vrid")
#> Warning: Targets for age_weight do not sum to 100%. Adjusting values to total 100%
#> [1] "Raking converged in 18 iterations"
svy <- svy_wts$svy
sapply(wtvars, function(x) weights::wpct(svy[[x]], svy$weight))
#> $age_weight
#>     18-34     35-54       55+ 
#> 0.3514503 0.3486601 0.2998896 
#> 
#> $income_weight
#>      0-25K     25-35K     35-50K     50-75K    75-100K   100-150K      150K+ 
#> 0.17353726 0.09405487 0.14064324 0.18922885 0.15252025 0.16103245 0.08898309 
#> 
#> $race_weight
#>                 White              Hispanic Not white or Hispanic 
#>            0.73559778            0.17719770            0.08720452 
#> 
summary(svy_wts$wts)
#> $convergence
#> [1] "Complete convergence was achieved after 18 iterations"
#> 
#> $base.weights
#> [1] "No Base Weights Were Used"
#> 
#> $raking.variables
#> [1] "age_weight"    "income_weight" "race_weight"  
#> 
#> $weight.summary
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>   0.441   0.721   1.000   1.000   1.161   2.227 
#> 
#> $selection.method
#> [1] "variable selection conducted using _pctlim_ - discrepancies selected using _total_."
#> 
#> $general.design.effect
#> [1] 1.146602
#> 
#> $age_weight
#>          Target Unweighted N Unweighted %     Wtd N     Wtd % Change in %
#> 18-34 0.3514503          327    0.2766497  415.4143 0.3514503  0.07480056
#> 35-54 0.3486601          348    0.2944162  412.1162 0.3486601  0.05424384
#> 55+   0.2998896          507    0.4289340  354.4695 0.2998896 -0.12904440
#> Total 1.0000000         1182    1.0000000 1182.0000 1.0000000  0.25808881
#>        Resid. Disc. Orig. Disc.
#> 18-34  1.110223e-16  0.07480056
#> 35-54  2.220446e-16  0.05424384
#> 55+   -3.330669e-16 -0.12904440
#> Total  6.661338e-16  0.25808881
#> 
#> $income_weight
#>              Target Unweighted N Unweighted %     Wtd N      Wtd %  Change in %
#> 0-25K    0.17353726          178   0.15071973  204.9944 0.17353726  0.022817533
#> 25-35K   0.09405487          151   0.12785775  111.1042 0.09405487 -0.033802878
#> 35-50K   0.14064324          169   0.14309907  166.1377 0.14064324 -0.002455831
#> 50-75K   0.18922885          242   0.20491109  223.5304 0.18922885 -0.015682244
#> 75-100K  0.15252025          178   0.15071973  180.1676 0.15252025  0.001800522
#> 100-150K 0.16103245          182   0.15410669  190.2228 0.16103245  0.006925756
#> 150K+    0.08898309           81   0.06858594  105.1131 0.08898309  0.020397142
#> Total    1.00000000         1181   1.00000000 1181.2703 1.00000000  0.103881906
#>           Resid. Disc.  Orig. Disc.
#> 0-25K     1.387779e-16  0.022817533
#> 25-35K    0.000000e+00 -0.033802878
#> 35-50K   -2.775558e-17 -0.002455831
#> 50-75K   -6.661338e-16 -0.015682244
#> 75-100K   6.383782e-16  0.001800522
#> 100-150K  2.775558e-17  0.006925756
#> 150K+     1.249001e-16  0.020397142
#> Total     1.623701e-15  0.103881906
#> 
#> $race_weight
#>                           Target Unweighted N Unweighted %     Wtd N      Wtd %
#> White                 0.73559778          973   0.82387807  868.9398 0.73559778
#> Hispanic              0.17719770          114   0.09652837  209.3184 0.17719770
#> Not white or Hispanic 0.08720452           94   0.07959356  103.0121 0.08720452
#> Total                 1.00000000         1181   1.00000000 1181.2703 1.00000000
#>                        Change in %  Resid. Disc.  Orig. Disc.
#> White                 -0.088280289  4.440892e-16 -0.088280289
#> Hispanic               0.080669338 -3.885781e-16  0.080669338
#> Not white or Hispanic  0.007610952 -2.775558e-17  0.007610952
#> Total                  0.176560579  8.604228e-16  0.176560579
#>